refactor: extract DatabaseMigrator from ModelData (Phase 2a, -233 lines)
Extracts the 220-line v42-v140 Realm migration block from
ModelData.tryInitializeDatabase() into a new dedicated
DatabaseMigrator class.

## Changes

- **New file**: Models/DatabaseMigrator.swift (~259 lines)
  - Owns the full schema v42-v140 migration history
  - Builds the full Realm.Configuration with shouldCompactOnLaunch
  - Handles the legacy document-directory → application-support move
  - Single public API: makeConfiguration(statusHandler:) throws -> Realm.Configuration

- **ModelData.tryInitializeDatabase** reduced to a 6-line thin wrapper:
  ```
  func tryInitializeDatabase(statusHandler: @escaping (String) -> Void) throws {
      ModelData.RealmSchemaVersion = UInt64(YabrAppInfo.shared.build) ?? 1
      let conf = try DatabaseMigrator().makeConfiguration(statusHandler: statusHandler)
      Realm.Configuration.defaultConfiguration = conf
      realmConf = conf
  }
  ```

- Registered DatabaseMigrator.swift in the Xcode project (app + test targets)

## Verification

- xcodebuild build: SUCCEEDED
- ModelData.swift: 764 → 531 lines (-233, -30%)
- All migration logic is byte-identical to the original (zero behavior change)
